==> 02_G0arrestInMalignantCells/01_QuiescenceScoring.R <==
library(biomaRt)
library(plyr)
library(GeoTcgaData)
library(data.table)
library(sva)
library(class)
library(ggplot2)
library(ggpubr)
library(curl)
library(DGEobj.utils)
library(Seurat)
library(GSVA)
library(dplyr)

project_dir <- "~/BreastCancerG0arrest/"
# dir.create("G0arrestInMalignantCells")
setwd("02_G0arrestInMalignantCells")
# dir.create("data")

# Identifying G0 arrested cells in malignant cells------------------------------
# Adopted from the methods by Wiecek et al. (2023), Genome Biology [PMID: 37221612]

## Read in the data and subset malignant cells----------------------------------
integrated <- readRDS(paste0(project_dir, "/00_IntegratingDatasets/data/integrated.malignancy.rds"))
epithelial <- subset(integrated, subset = malignancy == "malignant")
expr.data <- GetAssayData(epithelial, slot = "data", assay = "RNA")
expr.data <- as.data.frame(expr.data)

# Load up- and down-regulated genes for G0 arrest:
# These gene sets were generated by Wiecek et al. (2023), Genome Biology
load("data/upregulated_common.RData")
load("data/downregulated_common.RData")
gene_list <- list(upregulated_common, downregulated_common)

# Keep only variable genes:
keep <- apply(expr.data, 1, function(x) var(x, na.rm = T) > 0)
var.expr.data <- data.matrix(expr.data[keep, ])

# Calculate z-scores using the G0 arrest gene sets:
zscore <- gsva(var.expr.data, gene_list, method = "zscore", verbose = T, parallel.sz = 1)
zscore <- t(zscore)
zscore <- data.frame(zscore)
zscore$combined <- zscore$X1 - zscore$X2
saveRDS(zscore, file = "data/zscores.rds")

# Calculate cut-off values for stringent Quiescent/Proliferating labelling:
library(maxstat)
library(survival)
zscore$positive <- 0
zscore$positive[zscore$combined > 0] <- 1
zscore$negative <- 0
zscore$negative[zscore$combined < 0] <- 1
Quiescent.mod <- maxstat.test(Surv(combined, positive) ~ combined, data = zscore, smethod = "LogRank", pmethod = "Lau94", iscores = T)
Proliferating.mod <- maxstat.test(Surv(X2, negative) ~ combined, data = zscore, smethod = "LogRank", pmethod = "Lau94", iscores = T)
Quiescent.cutoff <- Quiescent.mod[["estimate"]][["estimated cutpoint"]]
Proliferating.cutoff <- Proliferating.mod[["estimate"]][["estimated cutpoint"]]

# Part of figure 2a:
pdf("figures/zscoreDistribution.pdf", width = 8, height = 4)
par(mfrow = c(1, 2))
hist(zscore$combined, xlab = "z-score", main = "Frequency")
hist(zscore$combined, xlab = "z-score", main = "Probability", probability = T)
lines(density(zscore$combined), col = "blue3", lty = 2, lwd = 2)
abline(v = Quiescent.cutoff, col = "red4", lty = 2, lwd = 2)
abline(v = Proliferating.cutoff, col = "red4", lty = 2, lwd = 2)
dev.off()

idx <- match(colnames(integrated), rownames(zscore))
integrated$QuiescenceScore <- zscore$combined[idx]
# Extended Data Fig. 2a: 
pdf("figures/QuiescenceScoreUMAP.pdf", width = 6, height = 6)
FeaturePlot(integrated, raster = F, features = "QuiescenceScore", order = T, pt.size = 0.1) + 
  NoAxes() + ggtitle("Quiescence Score") + coord_fixed() +
  scale_colour_gradient2(low = "#228833", mid = "#F7F7F7", midpoint = Quiescent.cutoff, high = "#AA3377", na.value = "grey")
dev.off()
# Discrete categories:
integrated$QuiescenceStatus <- ifelse(integrated$QuiescenceScore > Quiescent.cutoff & integrated$malignancy == "malignant", "Quiescent"
                                     ifelse(integrated$QuiescenceScore < Proliferating.cutoff & integrated$malignancy == "malignant", "Proliferating", "Slow-cycling"))

save(Quiescent.cutoff, file = "data/Quiescent.cutoff.RData")
save(Proliferating.cutoff, file = "data/Proliferating.cutoff.RData")
saveRDS(integrated, file = "data/integrated_with_quiescence.rds")

# Some metrics about the quiescence---------------------------------------------
malignant_cells <- subset(integrated, subset = malignancy == "malignant")
table(malignant_cells$patient, malignant_cells$QuiescenceStatus) %>% data.frame() -> patient_quiescence_information
colnames(patient_quiescence_information) <- c("Patient", "QuiescenceStatus", "Cells")
patient_quiescence_information <- patient_quiescence_information %>% group_by(Patient) %>% mutate(Percentage = 100 * Cells / sum(Cells)) %>% ungroup()

# Pie chart showing the distribution of quiescence status for each patient:
patient_quiescence_summary <- patient_quiescence_information %>% 
  group_by(Patient, QuiescenceStatus) %>%
  summarise(Percentage = sum(Percentage)) %>% ungroup()

# Figure 2c:
pdf("figures/quiescence_status_pie_chart.pdf", width = 10, height = 10)
ggplot(patient_quiescence_summary, aes(x = "", y = Percentage, fill = QuiescenceStatus)) +
  geom_bar(width = 1, stat = "identity") +
  coord_polar("y", start = 0) +
  facet_wrap(~ Patient) + # Facet by Patient
  theme_void() +
  theme(legend.position = "none") +
  scale_fill_manual("QuiescenceStatus", values = c("Proliferating" = "#1B7837", "Quiescent" = "#762A83", "Slow-cycling" = "#BBBBBB")) +
  labs(fill = "Quiescence Status")
dev.off()

table(malignant_cells$type, malignant_cells$QuiescenceStatus) %>% data.frame() -> type_quiescence_information
colnames(type_quiescence_information) <- c("Type", "QuiescenceStatus", "Cells")
type_quiescence_information <- type_quiescence_information %>% group_by(Type) %>% mutate(Percentage = 100 * Cells / sum(Cells)) %>% ungroup()

# Figure 2b:
pdf("figures/quiescence_status_per_type.pdf", width = 6, height = 4)
ggplot(type_quiescence_information, aes(fill = QuiescenceStatus, y = Percentage, x = Type)) +
  geom_bar(position = "stack", stat = "identity") + 
  theme_classic() +
  theme(axis.text.x = element_text(angle = 45, vjust = 0.5, hjust = 0.5)) + 
  labs(x = "Type", y = "Cell cycle status", title = "Cell cycle status per type") +
  scale_fill_manual("QuiescenceStatus", values = c("Proliferating" = "#1B7837", "Quiescent" = "#762A83", "Slow-cycling" = "#BBBBBB")) 
dev.off()

# Re-integration of subset of epithelial cells----------------------------------
## read in & prepare the data:
integrated$celltype <- as.character(integrated$celltype)
integrated$celltype[integrated$QuiescenceStatus == "Quiescent"] <- "G0 arrested"
integrated$celltype[integrated$QuiescenceStatus == "Proliferating"] <- "Fast-cycling"
integrated$celltype[integrated$QuiescenceStatus == "Slow-cycling"] <- "Slow-cycling"
Idents(integrated) <- integrated$celltype
integrated <- subset(integrated, subset = cellclass == "EPI")
integrated <- subset(integrated, subset = (disease %in% "cancer") & !(integrated$patient %in% c("Patient.12", "Patient.13", "Patient.14", "Patient.15")))
integrated <- DietSeurat(integrated, assays = "RNA")

# split datasets and process without integration
integrated[["RNA"]] <- split(integrated[["RNA"]], f = integrated$orig.ident)
integrated <- integrated %>%
  SCTransform(vars.to.regress = c("percent.mt", "celltype", "patient")) %>%
  RunPCA() %>%
  RunUMAP(dims = 1:30)

# integrate datasets
integrated <- IntegrateLayers(object = integrated, method = CCAIntegration, normalization.method = "SCT")
integrated <- FindNeighbors(integrated, reduction = "integrated.dr", dims = 1:30)
integrated <- FindClusters(integrated, resolution = 0.6)

integrated <- RunUMAP(integrated, dims = 1:30, reduction = "integrated.dr")

# Figure 1d:
pdf("umap_malignancy.pdf", width = 8, height = 4)
DimPlot(integrated, reduction = "umap", group.by = c("orig.ident", "malignancy"))
dev.off()

# Figure 2a:
pdf("figures/umap_quiescence_status.pdf", width = 6, height = 6)
FeaturePlot(integrated, raster = F, features = "QuiescenceStatus", order = T, pt.size = 0.1) + 
  NoAxes() +
  ggtitle("Quiescence Status") + 
  coord_fixed()
dev.off()

# Extended Data Fig. 2a:
pdf("figures/umap_quiescence_score.pdf", width = 6, height = 6)
FeaturePlot(integrated, raster = F, features = "QuiescenceScore", order = T, pt.size = 0.1) + 
  NoAxes() +
  ggtitle("Quiescence Score") + 
  coord_fixed() + 
  scale_colour_gradient2(low = "#228833", mid = "#F7F7F7", high = "#AA3377", na.value = "grey")
dev.off()

setwd(project_dir)
